add tests for sapPreloaderOutside
