+++ test/fixtures/spa-loader/nuxt.config.ts
@@ -1,6 +1,12 @@
 export default defineNuxtConfig({
-  ssr: false,
   devtools: { enabled: false },
   spaLoadingTemplate: true,
+  routeRules: {
+    '/spa': { ssr: false },
+    '/ssr': { ssr: true },
+  },
+  experimental: {
+    spaPreloaderOutside: false,
+  },
   compatibilityDate: '2024-06-28',
 })
