+++ test/spa-loader/spa-preloader-outside-enabled.test.ts
@@ -6,26 +6,42 @@ import { getBrowser, setup, url } from '@nuxt/test-utils'
 const isWebpack = process.env.TEST_BUILDER === 'webpack' || process.env.TEST_BUILDER === 'rspack'
 
 await setup({
-  rootDir: fileURLToPath(new URL('./fixtures/spa-loader', import.meta.url)),
+  rootDir: fileURLToPath(new URL('../fixtures/spa-loader', import.meta.url)),
   dev: process.env.TEST_ENV === 'dev',
   server: true,
   browser: true,
   setupTimeout: (isWindows ? 360 : 120) * 1000,
   nuxtConfig: {
     builder: isWebpack ? 'webpack' : 'vite',
-    ssr: false,
     spaLoadingTemplate: true,
+    experimental: {
+      spaPreloaderOutside: true,
+    },
   },
 })
 
-describe('spa-loader with SPA', () => {
+describe('spaPreloaderOutside flag is enabled', () => {
   it('should render spa-loader', async () => {
     const browser = await getBrowser()
     const page = await browser.newPage({})
-    await page.goto(url('/'), { waitUntil: 'domcontentloaded' })
-
+    await page.goto(url('/spa'), { waitUntil: 'domcontentloaded' })
     const loader = page.getByTestId('loader')
     expect(await loader.isVisible()).toBeTruthy()
+
+    const content = page.getByTestId('content')
+    await content.waitFor({ state: 'visible' })
+    expect(await loader.isHidden()).toBeTruthy()
+
+    await page.close()
+  }, 60_000)
+
+  it('should render content without spa-loader', async () => {
+    const browser = await getBrowser()
+    const page = await browser.newPage({})
+    await page.goto(url('/ssr'), { waitUntil: 'domcontentloaded' })
+
+    const loader = page.getByTestId('__nuxt-spa-loader')
+    expect(await loader.isHidden()).toBeTruthy()
 
     const content = page.getByTestId('content')
     await content.waitFor({ state: 'visible' })
